+++ src/filter/edge.rs
@@ -8,7 +8,7 @@ use crate::error::ImgProcResult;
 use crate::util::constants::{K_PREWITT_1D_VERT, K_PREWITT_1D_HORZ, K_SOBEL_1D_VERT, K_SOBEL_1D_HORZ, K_LAPLACIAN};
 
 /// Applies a separable derivative mask to a grayscale image
-pub fn derivative_mask(input: &Image<f64>, vert_kernel: &[f64], horz_kernel: &[f64]) -> ImgProcResult<Image<f64>> {
+pub fn derivative_mask(input: &Image<f32>, vert_kernel: &[f32], horz_kernel: &[f32]) -> ImgProcResult<Image<f32>> {
     error::check_grayscale(input)?;
 
     let img_x = filter::separable_filter(&input, &vert_kernel, &horz_kernel)?;
@@ -24,37 +24,37 @@ pub fn derivative_mask(input: &Image<f64>, vert_kernel: &[f64], horz_kernel: &[f
 }
 
 /// Applies the Prewitt operator to a grayscale image
-pub fn prewitt(input: &Image<f64>) -> ImgProcResult<Image<f64>> {
+pub fn prewitt(input: &Image<f32>) -> ImgProcResult<Image<f32>> {
     Ok(derivative_mask(input, &K_PREWITT_1D_VERT, &K_PREWITT_1D_HORZ)?)
 }
 
 /// Applies the Sobel operator to a grayscale image
-pub fn sobel(input: &Image<f64>) -> ImgProcResult<Image<f64>> {
+pub fn sobel(input: &Image<f32>) -> ImgProcResult<Image<f32>> {
     Ok(derivative_mask(input, &K_SOBEL_1D_VERT, &K_SOBEL_1D_HORZ)?)
 }
 
 /// Applies a Sobel operator with weight `weight` to a grayscale image
-pub fn sobel_weighted(input: &Image<f64>, weight: u32) -> ImgProcResult<Image<f64>> {
-    let vert_kernel = vec![1.0, weight as f64, 1.0];
+pub fn sobel_weighted(input: &Image<f32>, weight: u32) -> ImgProcResult<Image<f32>> {
+    let vert_kernel = vec![1.0, weight as f32, 1.0];
     Ok(derivative_mask(input, &vert_kernel, &K_SOBEL_1D_HORZ)?)
 }
 
 /// Applies the Laplacian operator to a grayscale image. Output contains positive
 /// and negative values - use [`normalize_laplacian()`](fn.normalize_laplacian.html) for visualization
-pub fn laplacian(input: &Image<f64>) -> ImgProcResult<Image<f64>> {
+pub fn laplacian(input: &Image<f32>) -> ImgProcResult<Image<f32>> {
     Ok(filter::unseparable_filter(input, &K_LAPLACIAN)?)
 }
 
 /// Applies the Laplacian of Gaussian operator using a `size x size` kernel to a grayscale image.
 /// Output contains positive and negative values - use
 /// [`normalize_laplacian()`](fn.normalize_laplacian.html) for visualization
-pub fn laplacian_of_gaussian(input: &Image<f64>, size: u32, sigma: f64) -> ImgProcResult<Image<f64>> {
+pub fn laplacian_of_gaussian(input: &Image<f32>, size: u32, sigma: f32) -> ImgProcResult<Image<f32>> {
     let kernel = util::generate_log_kernel(size, sigma)?;
     Ok(filter::unseparable_filter(input, &kernel)?)
 }
 
 /// Normalizes the result of a Laplacian or Laplacian of Gaussian operator to the range [0, 255]
-pub fn normalize_laplacian(input: &Image<f64>) -> ImgProcResult<Image<u8>> {
+pub fn normalize_laplacian(input: &Image<f32>) -> ImgProcResult<Image<u8>> {
     error::check_grayscale(input)?;
 
     let min = *input.data().iter().min_by(|x, y| x.partial_cmp(y).unwrap()).unwrap();
